+++ packages/analytics/nestjs/src/lib/interceptor/analytics.interceptor.ts
@@ -1,8 +1,8 @@
 import { type Maybe } from '@dereekb/util';
 import { type AnalyticsEventName } from '@dereekb/analytics';
-import { type CallHandler, type ExecutionContext, Injectable, type NestInterceptor, SetMetadata } from '@nestjs/common';
-import { type Reflector } from '@nestjs/core';
-import { type EventEmitter2 } from '@nestjs/event-emitter';
+import { type CallHandler, type ExecutionContext, Injectable, type NestInterceptor, SetMetadata, Inject } from '@nestjs/common';
+import { Reflector } from '@nestjs/core';
+import { EventEmitter2 } from '@nestjs/event-emitter';
 import { type Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
@@ -64,8 +64,8 @@ export const EmitAnalyticsEvent = <T>(config: AnalyticsEventInterceptorConfig<T>
 @Injectable()
 export class AnalyticsEventInterceptor<T = any> implements NestInterceptor {
   constructor(
-    readonly reflector: Reflector,
-    readonly eventEmitter: EventEmitter2
+    @Inject(Reflector) readonly reflector: Reflector,
+    @Inject(EventEmitter2) readonly eventEmitter: EventEmitter2
   ) {}
 
   /**