+++ packages/dbx-web/src/lib/button/icon/icon.button.module.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-deprecated -- this module wraps the deprecated component */
 import { NgModule } from '@angular/core';
 import { DbxIconButtonComponent } from './icon.button.component';