+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [13.6.14](https://github.com/dereekb/dbx-components/compare/v13.6.13-dev...v13.6.14) (2026-04-02)
+
+
+### Build System
+
+- lint fix ([46d9ad19](https://github.com/dereekb/dbx-components/commit/46d9ad1959b55508b7766cc097672883ee6b0c8e))
+
+
+### Code Refactoring
+
+- added @dereekb/nestjs/eslint sub-entry ([0ed211b0](https://github.com/dereekb/dbx-components/commit/0ed211b07a9ba61d0f758e186c0beb0e5021ffe0))
+- added sidenav position, displayMode, and css tokens ([301bd3d6](https://github.com/dereekb/dbx-components/commit/301bd3d60d7adc2fe291e18a79486f8f308ae4c9))
+- added type-only import detection to rule ([72f6ee3d](https://github.com/dereekb/dbx-components/commit/72f6ee3db8729f0d2348164dc545b895421a8b89))
+- arktype fixes ([a5a22a93](https://github.com/dereekb/dbx-components/commit/a5a22a939b63e009bd702e6f663f0946729ecb5d))
+
 ## [13.6.13](https://github.com/dereekb/dbx-components/compare/v13.6.12-dev...v13.6.13) (2026-04-01)